Another dependency the bbot missed!!!
BUG=44471
TEST=none

Review URL: http://codereview.chromium.org/2201001

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@48198 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: d265369f2b05102f44e88b3586562060973ecf48

+++ chrome/installer/util/l10n_string_util.cc
@@ -42,7 +42,7 @@ std::wstring GetSystemLanguage(const bool use_omaha_language) {
   }
   length = GetLocaleInfo(id, LOCALE_SISO639LANGNAME,
                          WriteInto(&language, length), length);
-  DCHECK(length == language.length() + 1);
+  DCHECK(length == static_cast<int>(language.length() + 1));
   StringToLowerASCII(&language);
 
   // Add the country if we need it.
@@ -51,7 +51,7 @@ std::wstring GetSystemLanguage(const bool use_omaha_language) {
   if (0 != length) {
     length = GetLocaleInfo(id, LOCALE_SISO3166CTRYNAME,
                            WriteInto(&country, length), length);
-    DCHECK(length == country.length() + 1);
+    DCHECK(length == static_cast<int>(country.length() + 1));
     StringToLowerASCII(&country);
     if (L"en" == language) {
       if (L"gb" == country) {
